move Mmap and SharedArrays to stdlib
