+++ doc/src/stdlib/.gitignore
@@ -1,2 +1,4 @@
 delimitedfiles.md
 test.md
+mmap.md
+sharedarrays.md
