+++ stdlib/Mmap/docs/src/index.md
@@ -1,0 +1,7 @@
+# Memory-mapped I/O
+
+```@docs
+Mmap.Anonymous
+Mmap.mmap
+Mmap.sync!
+```
