+++ stdlib/Mmap/test/runtests.jl
@@ -1,5 +1,7 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
+using Test, Mmap
+
 file = tempname()
 write(file, "Hello World\n")
 t = b"Hello World"
