+++ stdlib/SharedArrays/docs/src/index.md
@@ -1,0 +1,9 @@
+# Shared Arrays
+
+```@docs
+SharedArrays.SharedArray
+SharedArrays.procs(::SharedArray)
+SharedArrays.sdata
+SharedArrays.indexpids
+SharedArrays.localindexes
+```
