+++ test/compile.jl
@@ -212,7 +212,7 @@ try
                                     [:Base, :Core, Foo2_module, FooBase_module, :Main, :Test]),
                                # plus modules included in the system image
                                Dict(s => Base.module_uuid(Base.root_module(s)) for s in
-                                    [:DelimitedFiles]))
+                                    [:DelimitedFiles,:Mmap]))
         @test deps == deps1
 
         @test current_task()(0x01, 0x4000, 0x30031234) == 2
